# public/filash-openapi.yaml
openapi: 3.0.3
info:
  title: Filash Cloud File Server
  description: |-
    Filash cloud file server is the next generation file server specifically designed for the cloud era with native features built for hybrid cloud enviornments with both public and private cloud deployments as well as various on-premise storages such as hard drive and SSD. With the next generation features listed below, it is the best choice in any environment and the perfect replacement of any old servers.
    
    It overcomes the most challenging defects of the current file servers such as FTP, SFTP, scp and etc. With the state of the art design and implementation of data transportation aglorithms powered by AI, it triumphs by 
    
    * all data is encrypted with the most secure encryption algorithm
    
    * transfer any size data at full speed of the network/internet utilizing the full bandwidth with AI powered adaptive and adjustable rate control
    
    * high network delay doesn't affect the speed, cross continent, satellite earth, interplanetary transfer speed is the same as same city
    
    * high data loss rate doesn't affect the speed, with old file servers, over 10% loss rate will make them unworkable, while filash cloud file server still works fine with over 90% loss rate.
    
...
    
    * [find more features](https://filash.io)
    
  
  termsOfService: https://filash.io/terms/
  contact:
    email: filash@filash.io
  license:
    name: Proprietary
...
  version: 1.0.0
...
servers:
...
paths:
...
  /users:
    parameters:
      - $ref: '#/components/parameters/filash-domain'
    post:
      tags:
        - users
      summary: Create user in the givien domain
      description: This can only be done by the logged in user. The very first super user can be created by running the command line like `filash admin -o -d'{"id":"super","password":"1234"}`. If the domain field is empty, a random value will be generated. Users created by higher role users are forbidden to be updated or delete by lower role uers.
      operationId: createUser
      requestBody:
        description: The user object to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '409':
          $ref: '#/components/responses/409'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
    get:
      tags:
        - users
      summary: List users in the given domain
      description: Super user can list users of any domain, admin users can list users of their own domain, regular users have no permission to list.
      operationId: getUsers
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/tag'
        - $ref: '#/components/parameters/filash-next-token'
      responses:
        '200':
          description: successful operation
          headers:
            filash-next-token:
              $ref: '#/components/headers/filash-next-token'
            X-Count:
              $ref: '#/components/headers/X-Count'
            X-Total-Count:
              $ref: '#/components/headers/X-Total-Count'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/id'
    get:
      tags:
        - users
      summary: Get the details of an user
      description: Super user can get a user details of any domain, admin users can get a user details of their own domain, regular only gets own details.
      operationId: getUser
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
    put:
      tags:
        - users
      summary: Update an existing user
      description: Super user can update any user, admin user can update users of its own domain, regular user can only update self.
      operationId: putUser
      parameters:
        - $ref: '#/components/parameters/filash-settings-copy'
      requestBody:
        description: The details of the user to be updated.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          $ref: '#/components/responses/200'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
    delete:
      tags:
        - users
      summary: Delete an user
      description: Super user can delete any user, admin user can delete users of its own domain, regular user can only delete self, self can also be deleted.
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/last_updated_at'
      responses:
        '200':
          $ref: '#/components/responses/200'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
...
  /transfers:
    parameters:
      - $ref: '#/components/parameters/filash-domain'
    post:
      tags:
        - transfers
      summary: Create a transfer.
      description: |-
        This can only be done by the logged in user. Super users can create transfers in any domain, admin and regular users can create transfers in own domain only.
        
        Parallel session transfers can be created if allow_parallel is true in settings

        Transfers created by higher role users are forbidden to be updated or delete by lower role uers.
      operationId: createTransfer
      parameters:
        - $ref: '#/components/parameters/filash-follow'
      requestBody:
        description: The transfer object to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transfer'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '409':
          $ref: '#/components/responses/409'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
    get:
      tags:
        - transfers
      summary: List transfers in the given domain
      description: Super user can list transfers of any domain, admin users can list transfers of their own domain, regular users can list own transfers.
      operationId: getTransfers
      parameters:
        - $ref: '#/components/parameters/status'
        - $ref: '#/components/parameters/user'
        - $ref: '#/components/parameters/storage'
        - $ref: '#/components/parameters/share'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/tag'
        - $ref: '#/components/parameters/filash-next-token'
      responses:
        '200':
          description: successful operation
          headers:
            filash-next-token:
              $ref: '#/components/headers/filash-next-token'
            X-Count:
              $ref: '#/components/headers/X-Count'
            X-Total-Count:
              $ref: '#/components/headers/X-Total-Count'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transfer'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
  /transfers/{id}:
    parameters:
      - $ref: '#/components/parameters/filash-domain'
      - $ref: '#/components/parameters/id'
    get:
      tags:
        - transfers
      summary: Get the details of a transfer
      description: Super user can get a transfer details of any domain, admin users can get a transfer details of their own domain, regular users can get own transfer details.
      operationId: getTransfer
      parameters:
        - $ref: '#/components/parameters/filash-follow'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        '401':
          $ref: '#/components/responses/401'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
    put:
      tags:
        - transfers
      summary: Update/abort an existing transfer
      description: Super user can update a transfer of any domain, admin users can update a transfer of their own domain, regular users can update own transfers. Sync settings, resume and spec settings can be updated. The status of the transfer can be updated to 'enqueued' for rerun, to 'abort' to abort the active (enqueued or running) transfer.
      operationId: putTransfer
      parameters:
        - $ref: '#/components/parameters/filash-settings-copy'
      requestBody:
        description: The details of the transfer to be updated.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transfer'
      responses:
        '200':
          $ref: '#/components/responses/200'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
    delete:
      tags:
        - transfers
      summary: Delete a transfer
      description: Super users can delete any transfer, admin users can delete transfer of their own domain, regular users can delete own transfers.
      operationId: deleteTransfer
      parameters:
        - $ref: '#/components/parameters/last_updated_at'
      responses:
        '200':
          $ref: '#/components/responses/200'
        '400':
          $ref: '#/components/responses/400'
        '401':
          $ref: '#/components/responses/401'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
        '503':
          $ref: '#/components/responses/503'
...
components:
  schemas:
    UserSettings:
      type: object
      properties:
        max_number_of_authorized_keys_per_user:
          type: integer
          format: int64
          default: 1<<63 - 1
          example: 100
...
    User:
      type: object
      properties:
        id:
          type: string
          example: us.cgd9buo3lcj4di4njrtg
        password:
          type: string
          example: 12345
        salt:
          type: string
          example: cgda8cg3lcj5akmrd18g
        role:
          type: string
          example: regular
          enum:
            - regular
            - admin
            - super
        settings:
          type: object
          $ref: '#/components/schemas/UserSettings'
        tags:
          type: object
          additionalProperties: {}
          example: '{"org":{"id":"finance_1"}}'
        domain:
          type: string
          example: do.cgd9buo3lcj4di4njrug
        created_by:
          type: string
          example: admin
        created_at:
          type: string
          format: date-time
          description: in format of RFC3339
          example: 2023-03-01T12:45:31Z
        last_updated_by:
          type: string
          example: admin
        last_updated_at:
          type: string
          format: date-time
          description: in format of RFC3339
          example: 2023-03-01T12:45:31Z
...
    Transfer:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          example: enqueued
          enum:
            - enqueued
            - running
            - completed
        spec:
          type: object
          $ref: '#/components/schemas/TransferSpec'
        enqueued_sessions:
          type: integer
          format: int64
        running_sessions:
          type: integer
          format: int64
        completed_sessions:
          type: integer
          format: int64
        aborted_sessions:
          type: integer
          format: int64
        progress:
          type: object
          $ref: '#/components/schemas/Progress'
        retries:
          type: integer
          format: int64
          description: the number of retried times if the transfer failed
        error:
          type: string
        initiator:
          type: boolean
          description: indicating whether it is submitted on the current server
        tags:
          type: object
          additionalProperties: {}
          example: '{"org":{"id":"finance_1"}}'
        domain:
          type: string
          example: do.cgd9buo3lcj4di4njrug
        created_by:
          type: string
          example: admin
        created_at:
          type: string
          format: date-time
          description: in format of RFC3339
          example: 2023-03-01T12:45:31Z
        last_enqueued_at:
          type: string
          format: date-time
          description: in format of RFC3339
          example: 2023-03-01T12:45:31Z
        last_updated_by:
          type: string
          example: admin
        last_updated_at:
          type: string
          format: date-time
          description: in format of RFC3339
          example: 2023-03-01T12:45:31Z
...
    Progress:
      type: object
      properties:
        found_bytes:
          type: integer
          format: int64
        identical_bytes:
          type: integer
          format: int64
        expected_bytes:
          type: integer
          format: int64
        reloaded_bytes:
          type: integer
          format: int64
        read_bytes:
          type: integer
          format: int64
        assembled_bytes:
          type: integer
          format: int64
        rebuilt_bytes:
          type: integer
          format: int64
        written_bytes:
          type: integer
          format: int64
        tranferred_bytes:
          type: integer
          format: int64
        retransferred_bytes:
          type: integer
          format: int64
        found_files:
          type: integer
          format: int64
        identical_files:
          type: integer
          format: int64
        expected_files:
          type: integer
          format: int64
        written_files:
          type: integer
          format: int64
        found_links:
          type: integer
          format: int64
        expected_links:
          type: integer
          format: int64
        written_Links:
          type: integer
          format: int64
        found_folders:
          type: integer
          format: int64
        expected_folders:
          type: integer
          format: int64
        made_folders:
          type: integer
          format: int64
        deleted_files:
          type: integer
          format: int64
        rtt:
          type: string
          format: int64
          description: in unit of ms
          example: 100ms
        bps: 
          type: integer
          format: int64
          description: bits per second
        progress:
          type: number
          format: double
    TransferSpec:
      type: object
      properties:
        action:
          type: string
          example: upload
          enum:
            - upload
            - download
        sync:
          type: object
          properties:
            mode:
              type: string
              example: transfer
              enum:
                - transfer
                - session
            interval:
              type: string
              description: in such units as 's' as second,  'm' as minute, 'h' as hour
              example: 1h
            since_last_updated_at:
              type: boolean
              default: false
              description: if true, the interval will be calcuated by the last_updated_at, otherwise by last_enqueued_at
            until:
              type: string
              format: date-time
              description: in format of RFC3339
              example: 2023-04-11T16:00:25Z
        resume:
          type: string
          example: attr
          enum:
            - size
            - attr
            - partial
        source:
          type: object
          properties:
            share:
              type: string
            token:
              type: string
              example: Basic dXNlcjE6dXNlcjFwYXNzCg==
            cd:
              type: string
              description: current directory relative to the share root
            paths:
              type: array
              items:
                type: string
                example: ["file1.txt", "file2.txt"]
            symlink:
              type: string
              example: asis
              enum:
                - asis
                - follow
                - skip
            filters:
              type: object
              $ref: '#/components/schemas/Filters'
            remove_after:
              type: boolean
              default: false
              description: whether to delete files after being successfully transferred
        dest:
          type: object
          properties:
            share:
              type: string
            token:
              type: string
              example: Basic dXNlcjE6dXNlcjFwYXNzCg==
            path:
              type: string
        storage:
          type: string
          description: storage id of local side (source side for upload, dest side for download)
        remote_address:
          type: string
          description: comma seperated remote host names or ipv4/ipv6 addresses
          example: localhost
        path_pairs:
          type: array
          description: to transfer the full relative 'src' path to destination as the full relative path of 'dst'
          items:
            type: object
            properties:
              src:
                type: string
              dst:
                type: string
          example: [{"src":"folder_src_1/file1.txt","dst":"folder_dst_1/file1.txt"}]
        parallel:
          type: integer
          format: int64
          description: number of parallel sessions of the transfer, the source file set will be split among parallel sessions, and big files will be split too according to parallel_file_size_threshold in settings
        settings:
          $ref: '#/components/schemas/TransferSpecSettings'
...
    Filters:
      type: object
      properties:
        includes:
          type: array
          items:
            type: string
            example: "**.txt"
        excludes:
          type: array
          items:
            type: string
            example: .filash
        before:
          type: string
          format: date-time
          description: in format of RFC3339
          example: 2023-04-11T16:00:25Z
        after:
          type: string
          format: date-time
          description: in format of RFC3339
          example: 2023-04-11T16:00:25Z
...
    ResponseMessage:
      type: object
      properties:
        messsage:
          type: string
...
  parameters:
    id:
      name: id
      in: path
      description: unique ID of the resource
      required: true
      schema:
        type: string
...
    filash-domain:
      name: filash-domain
      in: header
      description: required for super user login to hint the domain
      schema:
        type: string
        example: do.cgd9buo3lcj4di4njrtf
    filash-settings-copy:
      name: filash-settings-copy
      in: header
      description: indicating whether or not copy the whole settings in the request body, which is used to unset settings, copy if present
      required: false
      schema:
        type: string
        example: yes
    last_updated_at:
      name: last_updated_at
      in: header
      description: it is required for confirmation to avoid chaos of concurrent modify requests, it can be obtained from the response of GET /{resource}/{id}
      required: true
      schema:
        type: string
...
    filash-follow:
      name: filash-follow
      in: header
      description: follow the transfer with http streaming synchronously, if present, transfer progress will be http streaming to the client until the status turns completed
      required: false
      schema:
        type: string
        example: yes
...
    storage:
      name: storage
      in: query
      description: scope the query with the storage id
      required: false
      schema:
        type: string
        example: st.cgd9buo3lcj4di4njrtf
    share:
      name: share
      in: query
      description: scope the query with the share id
      required: false
      schema:
        type: string
        example: sh.cgd9buo3lcj4di4njrtf
    status:
      name: status
      in: query
      description: active status jobs including enqueued and running can be queried with this parameter
      required: false
      schema:
        type: string
        example: running
        enum:
          - enqueued
          - running
    user:
      name: user
      in: query
      description: the user id to scope the query, regular user login ignores this parameter
      required: false
      schema:
        type: string
    limit:
      name: limit
      in: query
      description: the number of items in the response, not exceeding max_number_of_records_in_response in the config
      required: false
      schema:
        type: integer
        format: int64
        default: 50
    tag:
      name: tag
      in: query
      description: the key/value tag used to query, for example if the user has the tags json like '{"tags":{"org":{"id":"finance_1"}}}', the tag query parameter can be '?tag=org.id=finance_1'
      required: false
      schema:
        type: string
        example: org.id=finance_1
    filash-next-token:
      name: filash-next-token
      in: header
      schema:
        $ref: '#/components/headers/filash-next-token'
  responses:
    '200':
      description: successful operation
    '400':
      description: bad request for invalid parameters.
    '401':
      description: Authorization information is missing or invalid.
    '403':
      description: Forbidden for various reasons.
      content:
        application/json: 
          schema:
            $ref: '#/components/schemas/ResponseMessage'
    '404':
      description: Resource not found
    '409':
      description: Resource already exists
    '500':
      description: json parsing, database operations or other unexpected error.
      content:
        application/json: 
          schema:
            $ref: '#/components/schemas/ResponseMessage'
    '503':
      description: service unavailable
  headers:
    filash-next-token:
      description: a base64 encoded string from which the list starts, if none, it starts from the first, bad base64 format will get 400 response code.
      required: false
      schema:
        type: string
        example: L1NHVm9jazFhVlRJPQ==
    X-Count:
      description: number of records in the current response
      schema:
        type: integer
        format: int64
        example: 10
    X-Total-Count:
      description: total number of records of the query that can be traversed with filash-next-token headers until zero records returned in the response.
      schema:
        type: integer
        format: int64
        example: 100
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Custom JWT
      $ref: '#/components/schemas/BearerToken'
security:
  - basicAuth: []
  - bearerAuth: []